Changed transmit to used internal variable.
The state machine is OK but is very slow to react and caused numerous
niggly problems with lockups as state changes were being missed.

This commit fixes a problem with transmit that replaces some of the transmit
state machine checking with internal variables.  Now it is reliable so
the state machine has to go.